Start with the jobs view (WIP)
I need to stop with the jobs view to fix an urgent problem in the db
design: the providers and seekers all have zip and city, which
prohiobits implementing simple and optimized scope based findes in the
models and abilities. The idea is to replace it with the places we
already have for broker regions.

To be continued...