Start with subdomain and places.
This separates the site into two parts: The main site, which purpose is
solely to provide information about the page, and the regional sites
that runs in a subdomain.

Due to this, zip/city is also dropped from seekers, providers and
organizations and replaces with the newer places concept, which allows
us to have more accurate information about the users places, like geo
data, validation and consistent naming.

This implies that user registration is now only open to people living
in a region that smalljobs supports.