Remove Facebook OmniAuth.
Currently Facebook Omniauth has very little benefits but introduces
much complexity to the signup process: The only benefit so far is that
the user doesn't need to set a password, but it allows the
registration of invalid users that does not belong to a supported
region, because all we get from Facebook is the name and email. This
means to have a valid registration process (which we do not have at
this time), we need a 2 step registration: 1. Omniauth, 2. Profile
completion. At this stage we cannot guarantee the user is able to
register, because his region may be unsupported.

In order to not spend to much time is features with little benefit,
OAuth is postponed for now...